rename utils -> lib
