+++ src/shared/lib/consts/globalVariables.ts
