+++ src/shared/lib/hooks/useAppStore.ts
