+++ src/shared/lib/hooks/useAutoId.ts
