+++ src/shared/lib/hooks/useQueryParams.ts
@@ -1,5 +1,5 @@
 'use client';
-import { initCurrentPage, selectOptionPagination, useDebounce } from '@/shared/utils';
+import { initCurrentPage, selectOptionPagination, useDebounce } from '@/shared/lib';
 import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 
 export const useQueryParams = () => {
