+++ src/shared/lib/index.ts
@@ -1,3 +1,3 @@
 export * from './consts/globalVariables';
 export * from './hooks';
-export * from './locales/i18nConfiguration';
+export * from '@/shared/lib/locales/i18nConfiguration';
