+++ src/shared/lib/locales/languages/en/translation.json
