+++ src/shared/lib/locales/languages/kz/translation.json
