+++ src/shared/lib/locales/languages/ua/translation.json
