+++ src/shared/ui/alert/ui/Alert.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import CloseOutline from '@/../public/assets/components/CloseOutline';
+import { useActions, useAppSelector } from '@/shared/lib';
 import { Button, Typography } from '@/shared/ui';
 import { alertSelectors } from '@/shared/ui/alert/model/alertSlice';
-import { useActions, useAppSelector } from '@/shared/utils';
 import { clsx } from 'clsx';
 
 import s from './Alert.module.scss';
