+++ src/shared/ui/input/Input.tsx
@@ -5,8 +5,8 @@ import EyeOff from '@/../public/assets/components/EyeOff'; //'@/assets/icons/svg
 import Close from '@/../public/assets/components/Close';
 import Eye from '@/../public/assets/components/Eye';
 import Search from '@/../public/assets/components/Search';
+import { useAutoId } from '@/shared/lib';
 import { Button, Typography } from '@/shared/ui';
-import { useAutoId } from '@/shared/utils';
 import clsx from 'clsx';
 
 import s from './Input.module.scss';
