+++ src/widget/userCard/UserCard.tsx
@@ -4,8 +4,8 @@ import { useState } from 'react';
 import Block from '@/../public/assets/components/Block';
 import avatarMock from '@/../public/avatar-mock.jpg';
 import { appSlice } from '@/app/model/appSlice';
+import { ModalKey, useAppDispatch, useModal } from '@/shared/lib';
 import { Button, Typography } from '@/shared/ui';
-import { ModalKey, useAppDispatch, useModal } from '@/shared/utils';
 import { ViewPhotoModal } from '@/widget/viewPhotoModal/ViewPhotoModal';
 import Image from 'next/image';
