+++ src/widget/viewPhotoModal/ViewPhotoModal.tsx
@@ -8,8 +8,8 @@ import { AddComment } from '@/features/addComment/AddComment';
 import { MoreActions } from '@/features/moreActions/MoreActions';
 import { ShowLikers } from '@/features/showLikers/ShowLikers';
 import { ToggleLike } from '@/features/toggleLike/ToggleLike';
+import { ModalKey, useModal } from '@/shared/lib';
 import { Modal } from '@/shared/ui';
-import { ModalKey, useModal } from '@/shared/utils';
 import { PostInteractionBar } from '@/widget/postInteractionBar/PostInteractionBar';
 import { UsersLikesModal } from '@/widget/usersLikesModal/UsersLikesModal';
 import Image from 'next/image';
